+++ server/http/store/cloud-traffic-metering.integration.test.ts
@@ -1,12 +1,12 @@
 import { sql } from 'drizzle-orm'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-import { S3Service } from '../adapters/gateways/s3'
-import { createShareRepo } from '../adapters/repos/share'
-import { cloudTrafficReports } from '../db/schema'
-import { currentTrafficPeriod } from '../domain/quota'
-import type { Database } from '../platform/interface'
-import { authedHeaders, createTestApp, seedBusinessLicense } from '../test/setup'
-import { encodeChildRef } from './share-utils'
+import { S3Service } from '../../adapters/gateways/s3'
+import { createShareRepo } from '../../adapters/repos/share'
+import { cloudTrafficReports } from '../../db/schema'
+import { currentTrafficPeriod } from '../../domain/quota'
+import type { Database } from '../../platform/interface'
+import { authedHeaders, createTestApp, seedBusinessLicense } from '../../test/setup'
+import { encodeChildRef } from '../share-utils'
 
 const STORAGE_ID = 'st-cloud-traffic-test'
