+++ server/http/store/shared.ts
