+++ server/usecases/redirect.test.ts
@@ -1,5 +1,4 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { type DownloadTrafficOutcome, meterDownloadTraffic, reportDownloadEgress } from './cloud-traffic-metering'
 import type {
   ImageHostingRepo,
   ImageResolution,
@@ -13,14 +12,15 @@ import type {
   StorageRepo,
 } from './ports'
 import { type RedirectDeps, resolveDirectShareDownload, resolveImageHostingDownload } from './redirect'
+import { type DownloadTrafficOutcome, meterDownloadTraffic, reportDownloadEgress } from './store/cloud-traffic-metering'
 
 // The end-to-end metering (quota consume → cloud egress report → refund) is
 // covered by cloud-traffic-metering.test.ts. Here we replace its two entry
 // points with mocks so each redirect case feeds a chosen outcome and we can
 // assert the redirect flow's gates, ordering, and presign-rollback in isolation.
 // Everything else in the module (types, other exports) stays real.
-vi.mock('./cloud-traffic-metering', async (importOriginal) => ({
-  ...(await importOriginal<typeof import('./cloud-traffic-metering')>()),
+vi.mock('./store/cloud-traffic-metering', async (importOriginal) => ({
+  ...(await importOriginal<typeof import('./store/cloud-traffic-metering')>()),
   meterDownloadTraffic: vi.fn(),
   reportDownloadEgress: vi.fn(),
 }))
