+++ server/usecases/redirect.ts
@@ -8,9 +8,13 @@
 // inputs (cloud base URL, referer header, request origin), and renders the
 // route-specific Responses from the discriminated outcomes below.
 
-import { type CloudTrafficMeteringDeps, meterDownloadTraffic, reportDownloadEgress } from './cloud-traffic-metering'
 import type { ImageHostingRepo, QuotaRepo, S3Gateway, ShareRepo, StorageRepo } from './ports'
 import { PRESIGN_TTL_SECS } from './share'
+import {
+  type CloudTrafficMeteringDeps,
+  meterDownloadTraffic,
+  reportDownloadEgress,
+} from './store/cloud-traffic-metering'
 
 // The metering usecases need the cloud-report ports plus quota; the redirect
 // flows additionally read shares / image-hosting / storages and presign via s3.
