+++ server/usecases/store/cloud-store.test.ts
@@ -1,6 +1,8 @@
 import type { CloudOrderQuotaChange } from '@shared/schemas'
 import type { CloudStoreTarget } from '@shared/types'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
+import type { CloudStoreBinding, CloudStoreRepo, EffectiveQuota, LicensingCloudGateway, QuotaRepo } from '../ports'
+import { verifyCloudEventToken } from '../site/licensing'
 import {
   type CloudStoreDeps,
   cancelOrder,
@@ -15,13 +17,11 @@ import {
   processDeliveryWebhook,
   redeemGiftCard,
 } from './cloud-store'
-import type { CloudStoreBinding, CloudStoreRepo, EffectiveQuota, LicensingCloudGateway, QuotaRepo } from './ports'
-import { verifyCloudEventToken } from './site/licensing'
 
 // Token verification derives from a signed PASETO + trusted keys — out of scope
 // for a usecase unit test. Mock it so each case chooses verified/invalid; the
 // real path is covered by cloud-store.integration.test.ts.
-vi.mock('./site/licensing', () => ({ verifyCloudEventToken: vi.fn() }))
+vi.mock('../site/licensing', () => ({ verifyCloudEventToken: vi.fn() }))
 
 const verified = (eventId: string) => vi.mocked(verifyCloudEventToken).mockReturnValue({ eventId } as never)
 const rejected = () => vi.mocked(verifyCloudEventToken).mockReturnValue(null)
